zultra/handlers/inline.py
"""Inline query handlers for Zultra Telegram Bot."""

from telegram import Update
from telegram.ext import ContextTypes


async def handle_inline_query(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Handle inline queries."""
    # Inline query handling logic would go here
    pass