requirements.txt
python-telegram-bot[all]==20.8
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
redis==5.0.1
loguru==0.7.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==41.0.8
openai==1.6.1
google-generativeai==0.3.2
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
flake8==7.0.0